+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Fix mandatory feature template attributes when using device-specific variables
 - Add `sdwan_attach_feature_device_template` resource
 - Add `sdwan_vpn_list_policy_object` resource and data source
-- Add `sdwan_prefix_list_policy_object` resource and data source
+- Add `sdwan_ipv4_prefix_list_policy_object` resource and data source
 - Add `sdwan_policer_policy_object` resource and data source
 - Add `sdwan_mirror_policy_object` resource and data source
 - Add `sdwan_class_map_policy_object` resource and data source